Terminadas las clases sql de equipos, boletos y partidos

+++ BeiboletosJavaProject/Beiboletos/src/main/java/ModelosDAO/BoletoDAO.java
@@ -10,6 +10,6 @@ import ModelosD.Boleto;
  *
  * @author braul
  */
-public interface BoletoDAO extends DAO<Boleto, Long>{
+public interface BoletoDAO extends DAO<Boleto, Integer>{
     
 }
